backend/requirements.txt
fastapi
uvicorn
pillow
tflite-runtime
numpy
requests
